+++ R/NMsim.R
@@ -216,6 +216,7 @@
 ##'     you need to merge by a row identifier. You would do
 ##'     `args.NMscanData=list(col.row="ROW")` to merge by a column
 ##'     called `ROW`. This is only used in rare cases.
+##'
 ##' @param system.type A charachter string, either \"windows\" or
 ##'     \"linux\" - case insensitive. Windows is only experimentally
 ##'     supported. Default is to use \code{Sys.info()[["sysname"]]}.
@@ -408,6 +409,7 @@ NMsim <- function(file.mod,data,dir.sims, name.sim,
     ID <- NULL
     is.data <- NULL
     known <- NULL
+    MDV <- NULL
     model <- NULL
     mod <- NULL
     name.mod <- NULL