+++ R/getSystemType.R
@@ -1,11 +1,11 @@
 ##' @keywords internal
 
-getSystemType <- function(system.type){
+getSystemType <- function(sysname){
 
-    if(missing(system.type) || is.null(system.type)){
-        system.type <- Sys.info()['sysname']
+    if(missing(sysname) || is.null(sysname)){
+        sysname <- Sys.info()[['sysname']]
     }
-    system.type <- tolower(system.type)
+    system.type <- tolower(sysname)
     if(system.type=="darwin") system.type <- "linux"
     if(!system.type%in%c("linux","windows")){
         stop("system.type must be either linux or windows")